Refactor app and storage internals
