+++ src/app/service.rs
@@ -1,0 +1,420 @@
+use anyhow::{Context, Result};
+use uuid::Uuid;
+
+use crate::codex;
+use crate::env::AppEnv;
+use crate::model::{
+    ActivateOutput, DeleteOutput, DisplayIdentity, ListOutput, RunningCodexProcess, SaveAction,
+    SaveOutput, SnapshotBlob, StatusOutput,
+};
+use crate::repository::SnapshotRepository;
+use crate::secrets::SecretStore;
+
+use super::{
+    App, account_view, match_saved_account, saved_identity, should_verify_activation_stability,
+    subject_bound_identity_matches,
+};
+
+impl<S> App<S>
+where
+    S: SecretStore,
+{
+    pub fn new(env: AppEnv, repository: SnapshotRepository<S>) -> Self {
+        Self { env, repository }
+    }
+
+    pub fn status(&self) -> Result<StatusOutput> {
+        let saved_accounts = self.repository.list_accounts(&self.env.kind)?;
+        let live = codex::try_read_live_auth_bundle(&self.env)?;
+        let current_saved_id = live
+            .as_ref()
+            .and_then(|bundle| match_saved_account(&saved_accounts, &bundle.identity))
+            .map(|account| account.id);
+        Ok(StatusOutput {
+            environment: self.env.kind.clone(),
+            codex_root: self.env.codex_root.display().to_string(),
+            current_account: live.map(|bundle| bundle.identity),
+            current_account_saved_id: current_saved_id,
+            saved_accounts: saved_accounts.len(),
+            process_warnings: crate::process::detect_running_codex_processes(),
+        })
+    }
+
+    pub fn list(&self) -> Result<ListOutput> {
+        let accounts = self.repository.list_accounts(&self.env.kind)?;
+        let live = codex::try_read_live_auth_bundle(&self.env)?;
+        let active_id = live
+            .as_ref()
+            .and_then(|bundle| match_saved_account(&accounts, &bundle.identity))
+            .map(|account| account.id);
+        Ok(ListOutput {
+            environment: self.env.kind.clone(),
+            accounts: accounts
+                .into_iter()
+                .map(|account| account_view(account, active_id))
+                .collect(),
+        })
+    }
+
+    pub fn save_current(&self) -> Result<SaveOutput> {
+        let live = codex::read_live_auth_bundle(&self.env).with_context(|| {
+            format!(
+                "no live Codex auth bundle found at {}",
+                self.env.codex_root.display()
+            )
+        })?;
+        let (metadata, created) =
+            self.repository
+                .save_snapshot(&self.env.kind, &live.identity, &live.snapshot)?;
+        Ok(SaveOutput {
+            account: account_view(metadata.clone(), Some(metadata.id)),
+            action: if created {
+                SaveAction::Created
+            } else {
+                SaveAction::Refreshed
+            },
+        })
+    }
+
+    pub fn activate(&self, account_id: Uuid) -> Result<ActivateOutput> {
+        self.activate_with_running_policy(account_id, false)
+    }
+
+    pub fn validate_activation_target(&self, account_id: Uuid) -> Result<()> {
+        let _ = self.load_activation_target(account_id)?;
+        Ok(())
+    }
+
+    pub fn activate_with_running_policy(
+        &self,
+        account_id: Uuid,
+        force_running: bool,
+    ) -> Result<ActivateOutput> {
+        let warnings = crate::process::detect_running_codex_processes();
+        let (snapshot, snapshot_identity, restore_identity) =
+            self.load_activation_target(account_id)?;
+        let verify_stable = should_verify_activation_stability(force_running, &warnings);
+        codex::restore_snapshot(&self.env, &snapshot, &restore_identity, verify_stable)
+            .context("failed to restore the selected account snapshot")?;
+        let metadata = self
+            .repository
+            .sync_activated_account(&self.env.kind, account_id, &snapshot_identity)
+            .context("activated live auth but failed to update local metadata")?;
+        Ok(ActivateOutput {
+            account: account_view(metadata, Some(account_id)),
+            warnings,
+        })
+    }
+
+    fn load_activation_target(
+        &self,
+        account_id: Uuid,
+    ) -> Result<(SnapshotBlob, DisplayIdentity, DisplayIdentity)> {
+        let (metadata, snapshot) = self.repository.load_snapshot(&self.env.kind, account_id)?;
+        let expected_identity = saved_identity(&metadata);
+        let snapshot_identity = codex::identity_from_snapshot(&snapshot)?;
+        let restore_identity = if expected_identity.subject.is_some() {
+            if !subject_bound_identity_matches(&expected_identity, &snapshot_identity) {
+                anyhow::bail!(
+                    "saved snapshot identity does not match the selected account: expected {:?}, got {:?}",
+                    expected_identity,
+                    snapshot_identity
+                );
+            }
+            expected_identity.clone()
+        } else {
+            snapshot_identity.clone()
+        };
+        Ok((snapshot, snapshot_identity, restore_identity))
+    }
+
+    pub fn activation_preflight_warnings(&self) -> Vec<RunningCodexProcess> {
+        crate::process::detect_running_codex_processes()
+    }
+
+    pub fn delete(&self, account_id: Uuid) -> Result<DeleteOutput> {
+        self.repository
+            .delete_snapshot(&self.env.kind, account_id)?;
+        Ok(DeleteOutput {
+            deleted_account_id: account_id,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use base64::Engine;
+    use tempfile::tempdir;
+
+    use super::*;
+    use crate::codex::auth_json_fixture;
+    use crate::model::SnapshotFile;
+    use crate::model::{DisplayIdentity, EnvironmentKind};
+    use crate::secrets::test_support::MemorySecretStore;
+
+    #[test]
+    fn list_marks_active_account() {
+        let temp = tempdir().expect("tempdir");
+        let env = AppEnv {
+            kind: EnvironmentKind::Linux,
+            home_dir: temp.path().to_path_buf(),
+            codex_root: temp.path().join(".codex"),
+            app_data_dir: temp.path().join("app"),
+        };
+        std::fs::create_dir_all(&env.codex_root).expect("codex root");
+        std::fs::write(
+            env.codex_root.join("auth.json"),
+            auth_json_fixture("active@example.com", "sub-1", Some("pro")),
+        )
+        .expect("auth");
+        std::fs::write(env.codex_root.join("cap_sid"), "sid").expect("cap");
+        let repo = SnapshotRepository::new(&env.app_data_dir, MemorySecretStore::default());
+        repo.save_snapshot(
+            &env.kind,
+            &DisplayIdentity {
+                email: "active@example.com".to_owned(),
+                subject: Some("sub-1".to_owned()),
+                name: None,
+                plan_label: Some("Pro".to_owned()),
+            },
+            &SnapshotBlob {
+                schema_version: 1,
+                files: vec![],
+            },
+        )
+        .expect("save");
+        let app = App::new(env, repo);
+        let output = app.list().expect("list");
+        assert_eq!(output.accounts.len(), 1);
+        assert!(output.accounts[0].is_active);
+    }
+
+    #[test]
+    fn list_keeps_saved_account_when_live_account_is_unsaved() {
+        let temp = tempdir().expect("tempdir");
+        let env = AppEnv {
+            kind: EnvironmentKind::Linux,
+            home_dir: temp.path().to_path_buf(),
+            codex_root: temp.path().join(".codex"),
+            app_data_dir: temp.path().join("app"),
+        };
+        std::fs::create_dir_all(&env.codex_root).expect("codex root");
+        std::fs::write(
+            env.codex_root.join("auth.json"),
+            auth_json_fixture("current@example.com", "sub-2", Some("plus")),
+        )
+        .expect("auth");
+        std::fs::write(env.codex_root.join("cap_sid"), "sid-current").expect("cap");
+        let repo = SnapshotRepository::new(&env.app_data_dir, MemorySecretStore::default());
+        repo.save_snapshot(
+            &env.kind,
+            &DisplayIdentity {
+                email: "saved@example.com".to_owned(),
+                subject: Some("sub-1".to_owned()),
+                name: None,
+                plan_label: Some("Pro".to_owned()),
+            },
+            &SnapshotBlob {
+                schema_version: 1,
+                files: vec![],
+            },
+        )
+        .expect("save");
+        let app = App::new(env, repo);
+        let output = app.list().expect("list");
+        assert_eq!(output.accounts.len(), 1);
+        assert_eq!(output.accounts[0].email, "saved@example.com");
+        assert!(!output.accounts[0].is_active);
+    }
+
+    #[test]
+    fn subject_bound_identity_requires_matching_subject() {
+        let expected = DisplayIdentity {
+            email: "person@example.com".to_owned(),
+            subject: Some("sub-1".to_owned()),
+            name: Some("Tester".to_owned()),
+            plan_label: Some("Pro".to_owned()),
+        };
+        let missing_subject = DisplayIdentity {
+            email: "person@example.com".to_owned(),
+            subject: None,
+            name: Some("Tester".to_owned()),
+            plan_label: Some("Pro".to_owned()),
+        };
+        let wrong_subject = DisplayIdentity {
+            email: "person@example.com".to_owned(),
+            subject: Some("sub-2".to_owned()),
+            name: Some("Tester".to_owned()),
+            plan_label: Some("Pro".to_owned()),
+        };
+        let matching_subject = DisplayIdentity {
+            email: "other@example.com".to_owned(),
+            subject: Some("sub-1".to_owned()),
+            name: Some("Tester".to_owned()),
+            plan_label: Some("Pro".to_owned()),
+        };
+        assert!(!subject_bound_identity_matches(&expected, &missing_subject));
+        assert!(!subject_bound_identity_matches(&expected, &wrong_subject));
+        assert!(subject_bound_identity_matches(&expected, &matching_subject));
+    }
+
+    #[test]
+    fn activate_returns_refreshed_identity_after_subject_stable_restore() {
+        let temp = tempdir().expect("tempdir");
+        let env = AppEnv {
+            kind: EnvironmentKind::Linux,
+            home_dir: temp.path().to_path_buf(),
+            codex_root: temp.path().join(".codex"),
+            app_data_dir: temp.path().join("app"),
+        };
+        std::fs::create_dir_all(&env.codex_root).expect("codex root");
+        std::fs::write(
+            env.codex_root.join("auth.json"),
+            auth_json_fixture("current@example.com", "sub-current", Some("pro")),
+        )
+        .expect("auth");
+        std::fs::write(env.codex_root.join("cap_sid"), "sid-a").expect("cap");
+
+        let repo = SnapshotRepository::new(&env.app_data_dir, MemorySecretStore::default());
+        let saved = repo
+            .save_snapshot(
+                &env.kind,
+                &DisplayIdentity {
+                    email: "before@example.com".to_owned(),
+                    subject: Some("sub-1".to_owned()),
+                    name: Some("Before".to_owned()),
+                    plan_label: Some("Pro".to_owned()),
+                },
+                &SnapshotBlob {
+                    schema_version: 1,
+                    files: vec![
+                        SnapshotFile {
+                            name: "auth.json".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode(
+                                auth_json_fixture("after@example.com", "sub-1", Some("plus")),
+                            ),
+                        },
+                        SnapshotFile {
+                            name: "cap_sid".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode("sid-b"),
+                        },
+                    ],
+                },
+            )
+            .expect("save")
+            .0;
+
+        let app = App::new(env.clone(), repo);
+        let output = app.activate(saved.id).expect("activate");
+        assert_eq!(output.account.email, "after@example.com");
+        assert_eq!(output.account.plan_label.as_deref(), Some("Plus"));
+
+        let list = app.list().expect("list");
+        assert_eq!(list.accounts[0].email, "after@example.com");
+    }
+
+    #[test]
+    fn activate_rejects_snapshot_that_does_not_match_selected_account() {
+        let temp = tempdir().expect("tempdir");
+        let env = AppEnv {
+            kind: EnvironmentKind::Linux,
+            home_dir: temp.path().to_path_buf(),
+            codex_root: temp.path().join(".codex"),
+            app_data_dir: temp.path().join("app"),
+        };
+        std::fs::create_dir_all(&env.codex_root).expect("codex root");
+        std::fs::write(
+            env.codex_root.join("auth.json"),
+            auth_json_fixture("active@example.com", "sub-1", Some("pro")),
+        )
+        .expect("auth");
+        std::fs::write(env.codex_root.join("cap_sid"), "sid-a").expect("cap");
+
+        let repo = SnapshotRepository::new(&env.app_data_dir, MemorySecretStore::default());
+        let saved = repo
+            .save_snapshot(
+                &env.kind,
+                &DisplayIdentity {
+                    email: "expected@example.com".to_owned(),
+                    subject: Some("sub-expected".to_owned()),
+                    name: Some("Expected".to_owned()),
+                    plan_label: Some("Pro".to_owned()),
+                },
+                &SnapshotBlob {
+                    schema_version: 1,
+                    files: vec![
+                        SnapshotFile {
+                            name: "auth.json".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode(
+                                auth_json_fixture("wrong@example.com", "sub-wrong", Some("plus")),
+                            ),
+                        },
+                        SnapshotFile {
+                            name: "cap_sid".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode("sid-b"),
+                        },
+                    ],
+                },
+            )
+            .expect("save")
+            .0;
+
+        let app = App::new(env.clone(), repo);
+        let error = app.activate(saved.id).expect_err("activate should fail");
+        assert!(format!("{error:#}").contains("does not match the selected account"));
+        let live = crate::codex::read_live_auth_bundle(&env).expect("live bundle");
+        assert_eq!(live.identity.email, "active@example.com");
+    }
+
+    #[test]
+    fn activate_allows_legacy_metadata_without_subject_to_refresh_email() {
+        let temp = tempdir().expect("tempdir");
+        let env = AppEnv {
+            kind: EnvironmentKind::Linux,
+            home_dir: temp.path().to_path_buf(),
+            codex_root: temp.path().join(".codex"),
+            app_data_dir: temp.path().join("app"),
+        };
+        std::fs::create_dir_all(&env.codex_root).expect("codex root");
+        std::fs::write(
+            env.codex_root.join("auth.json"),
+            auth_json_fixture("active@example.com", "sub-1", Some("pro")),
+        )
+        .expect("auth");
+        std::fs::write(env.codex_root.join("cap_sid"), "sid-a").expect("cap");
+
+        let repo = SnapshotRepository::new(&env.app_data_dir, MemorySecretStore::default());
+        let saved = repo
+            .save_snapshot(
+                &env.kind,
+                &DisplayIdentity {
+                    email: "old@example.com".to_owned(),
+                    subject: None,
+                    name: Some("Old".to_owned()),
+                    plan_label: Some("Pro".to_owned()),
+                },
+                &SnapshotBlob {
+                    schema_version: 1,
+                    files: vec![
+                        SnapshotFile {
+                            name: "auth.json".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode(
+                                auth_json_fixture("new@example.com", "sub-new", Some("plus")),
+                            ),
+                        },
+                        SnapshotFile {
+                            name: "cap_sid".to_owned(),
+                            bytes_base64: base64::engine::general_purpose::STANDARD.encode("sid-b"),
+                        },
+                    ],
+                },
+            )
+            .expect("save")
+            .0;
+
+        let app = App::new(env.clone(), repo);
+        let output = app.activate(saved.id).expect("activate");
+        assert_eq!(output.account.email, "new@example.com");
+        assert_eq!(output.account.subject.as_deref(), Some("sub-new"));
+    }
+}
